Refactoring work done: Functional components created for camera and rover drop down.
