+++ mars-app-spa/src/selectedRoverContext.tsx
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+import { CameraI, RoverFullI } from "./nasaInterfaces";
+import { UpdatableRover } from "./UpdatableRover";
+
+export const selectedRoverContext = createContext<UpdatableRover>({
+  setRover: (rover: RoverFullI) => { },
+  setCamera: (camera: CameraI) => { },
+});
